Add reservation middleware (#193)
* Moved reservation code to the api package to make it easier to test and also there was no real point of separating it  
* Added middleware that checks for X-GO-IOS-RESERVE header for endpoints on which it is applied  
* Added a reservation admin UUID  
* Reworked reservation endpoints to be more in line with the other device endpoints  
* Added tests for the middleware and also cleaned up reservation tests code a bit
